Split out GlAnimation out of DataAnimation
